Move auction table to components, extract data for prebid table, and use auction table component
