+++ packages/extension/src/view/devtools/pages/privateAdvertising/protectedAudience/auctions/v1/panel/index.tsx
@@ -36,7 +36,7 @@ import { isEqual } from 'lodash-es';
  * Internal dependencies.
  */
 import type { AuctionEventsType } from '../../../../../../stateProviders/protectedAudience/context';
-import AuctionTable from '../table';
+import AuctionTable from '../../components/table';
 import AdunitPanel from '../adunitPanel';
 import AdunitSubPanel from '../adunitPanel/panel';
 import SortButton from '../../../../../sortButton';
